# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample app server.",
        "title": "Party Cost Calculator API",
        "contact": {},
        "version": "1.0"
    },
    "host": "s.cfg.Host:s.cfg.Port",
    "basePath": "/",
    "paths": {
...
        "/persEvents": {
            "get": {
                "description": "Get a record of peson existed in event by Id from database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Person-Event"
                ],
                "summary": "Get a person-event",
                "parameters": [
                    {
                        "description": "Get Person-Event Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/personsevents.GetPersonEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.PersonsAndEvents"
                        }
                    },
                    "304": {
                        "description": "Not Modified",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponce"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponce"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a record of peson existed in event by Id from database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Person-Event"
                ],
                "summary": "Update a person-event",
                "parameters": [
                    {
                        "description": "Update Person-Event Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/personsevents.UpdatePersonEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "304": {
                        "description": "Not Modified",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponce"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponce"
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new record of peson existed in event by Id to database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Person-Event"
                ],
                "summary": "Add a person-event",
                "parameters": [
                    {
                        "description": "Add Person-Event Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/personsevents.AddPersonEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "304": {
                        "description": "Not Modified",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponce"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponce"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a record of peson existed in event by Id from database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Person-Event"
                ],
                "summary": "Delete a person-event",
                "parameters": [
                    {
                        "description": "Delete Person-Event Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/personsevents.DeletePersonEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "304": {
                        "description": "Not Modified",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponce"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponce"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Event": {
            "type": "object",
            "properties": {
                "date": {
                    "description": "postgres format: `YYYY-MM-DD`",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "persons": {
                    "description": "PersonIds []int64",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Person"
                    }
                }
            }
        },
        "domain.Person": {
            "type": "object",
            "properties": {
                "events": {
                    "description": "EventIds []int64",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Event"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "domain.PersonsAndEvents": {
            "type": "object",
            "properties": {
                "event": {
                    "$ref": "#/definitions/domain.Event"
                },
                "eventId": {
                    "type": "integer"
                },
                "factor": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "person": {
                    "$ref": "#/definitions/domain.Person"
                },
                "personId": {
                    "type": "integer"
                },
                "spent": {
                    "type": "number"
                }
            }
        },
...
        "handlers.ErrorResponce": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "personsevents.AddPersonEventRequest": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "integer"
                },
                "factor": {
                    "type": "integer"
                },
                "person_id": {
                    "type": "integer"
                },
                "spent": {
                    "type": "number"
                }
            }
        },
        "personsevents.DeletePersonEventRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "personsevents.GetPersonEventRequest": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "integer"
                },
                "person_id": {
                    "type": "integer"
                }
            }
        },
        "personsevents.UpdatePersonEventRequest": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "integer"
                },
                "factor": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "person_id": {
                    "type": "integer"
                },
                "spent": {
                    "type": "number"
                }
            }
        },
...
    }
}